Conflicts:
	src/view/item/BasicItemView.java

+++ src/model/area/TileCoordinate.java
@@ -48,8 +48,8 @@ public class TileCoordinate implements Saveable {
 
 	public static RealCoordinate convertToRealCoordinate(TileCoordinate coord) {
 		double x = (0.75 * coord.getX());
-		double y = (double) ((Math.sqrt(3) / 2 * (double) coord.getY()) + (((double) coord
-				.getX() % 2) * Math.sqrt(3.0) / 4.0));
+		double y = (Math.sqrt(3) / 2 * coord.getY()) + (((double) coord
+				.getX() % 2) * Math.sqrt(3.0) / 4.0);
 		return new RealCoordinate(x, y);
 	}
